feat: interop rust to kotlin

+++ tyme4rs_clib/src/lib.rs
@@ -1,8 +1,10 @@
+use std::ffi::c_int;
+
 #[no_mangle]
-pub extern "C" fn add(left: u64, right: u64) -> u64 {
+pub extern "C" fn add(left: c_int, right: c_int) -> c_int {
     return add_impl(left, right);
 }
 
-pub fn add_impl(left: u64, right: u64) -> u64 {
+pub fn add_impl(left: c_int, right: c_int) -> c_int {
     left + right
 }
